rendering story images differently based on aspect ratio
Todo: Remove title from story.
Most important information in the stories is the descriptions and the even though the image speaks a lot, cropping some portions
of the image should be OK. Also the stories are designed to be consumed on mobile in 'cooking mode'. So optimize the image
for mobile screens and its ok to have margins and gaps in desktop version.

#9